requirements.txt
aiohappyeyeballs                        2.6.1
aiohttp                                 3.13.5
aiosignal                               1.4.0
anitya_schema                           2.3.0
annotated-doc                           0.0.4
annotated-types                         0.7.0
anyio                                   4.13.0
argcomplete                             3.6.3
async-timeout                           5.0.1
attrs                                   26.1.0
Automat                                 25.4.16
bleach                                  6.3.0
blinker                                 1.9.0
certifi                                 2026.2.25
cffi                                    2.0.0
charset-normalizer                      3.4.7
click                                   8.3.2
colorama                                0.4.6
constantly                              23.10.4
crochet                                 2.1.1
cryptography                            46.0.7
dataclasses-json                        0.6.7
decorator                               5.2.1
defusedxml                              0.7.1
distro                                  1.9.0
et_xmlfile                              2.0.0
fedora_messaging                        3.9.0
fedora_messaging_the_new_hotness_schema 1.3.0
filelock                                3.25.2
flatbuffers                             25.12.19
frozenlist                              1.8.0
fsspec                                  2026.3.0
future                                  1.0.0
greenlet                                3.3.2
groq                                    0.37.1
grpcio                                  1.80.0
gssapi                                  1.11.1
h11                                     0.16.0
h2                                      4.3.0
h5py                                    3.16.0
hf-xet                                  1.4.3
hpack                                   4.1.0
httpcore                                1.0.9
httpx                                   0.28.1
httpx-sse                               0.4.3
huggingface_hub                         1.9.1
hyperframe                              6.1.0
hyperlink                               21.0.0
idna                                    3.11
Incremental                             24.11.0
Jinja2                                  3.1.6
joblib                                  1.5.3
jsonpatch                               1.33
jsonpointer                             3.1.1
jsonschema                              4.26.0
jsonschema-specifications               2025.9.1
kaggle                                  2.0.0
kagglesdk                               0.1.16
koji                                    1.36.0
langchain                               1.2.15
langchain-classic                       1.0.3
langchain-community                     0.4.1
langchain-core                          1.2.27
langchain-groq                          1.1.2
langchain-huggingface                   1.2.1
langchain-text-splitters                1.1.1
langgraph                               1.1.6
langgraph-checkpoint                    4.0.1
langgraph-prebuilt                      1.0.9
langgraph-sdk                           0.3.13
langsmith                               0.7.26
lxml                                    6.0.2
markdown-it-py                          4.0.0
MarkupSafe                              3.0.3
marshmallow                             3.26.2
mdurl                                   0.1.2
mpmath                                  1.3.0
multidict                               6.7.1
mypy_extensions                         1.1.0
networkx                                3.6.1
numpy                                   2.4.4
onnxruntime                             1.24.4
openpyxl                                3.1.5
orjson                                  3.11.8
ormsgpack                               1.12.2
packaging                               26.0
pandas                                  3.0.2
pandas-stubs                            3.0.0.260204
pika                                    1.3.2
pillow                                  12.2.0
pip                                     26.0.1
portalocker                             3.2.0
propcache                               0.4.1
protobuf                                7.34.1
pyasn1                                  0.6.3
pyasn1_modules                          0.4.2
pycparser                               3.0
pydantic                                2.12.5
pydantic_core                           2.41.5
pydantic-settings                       2.13.1
Pygments                                2.20.0
PyMuPDF                                 1.27.2.2
pymupdf-layout                          1.27.2.2
pymupdf4llm                             1.27.2.2
pyOpenSSL                               26.0.0
python-bugzilla                         3.3.0
python-dateutil                         2.9.0.post0
python-docx                             1.2.0
python-dotenv                           1.2.2
python-slugify                          8.0.4
pywin32                                 311
PyYAML                                  6.0.3
qdrant-client                           1.17.1
rank-bm25                               0.2.2
redis                                   7.4.0
referencing                             0.37.0
regex                                   2026.4.4
requests                                2.33.1
requests-gssapi                         1.4.0
requests-toolbelt                       1.0.0
retriever                               3.1.0
rich                                    14.3.3
rpds-py                                 0.30.0
safetensors                             0.7.0
scikit-learn                            1.8.0
scipy                                   1.17.1
sentence-transformers                   5.3.0
service-identity                        24.2.0
setuptools                              80.3.1
shellingham                             1.5.4
six                                     1.17.0
sniffio                                 1.3.1
SQLAlchemy                              2.0.49
sympy                                   1.14.0
tabulate                                0.10.0
tenacity                                9.1.4
text-unidecode                          1.3
the-new-hotness                         1.4.1
threadpoolctl                           3.6.0
tokenizers                              0.22.2
tomli                                   2.4.1
torch                                   2.11.0
tqdm                                    4.67.3
transformers                            5.5.0
Twisted                                 25.5.0
typer                                   0.24.1
typing_extensions                       4.15.0
typing-inspect                          0.9.0
typing-inspection                       0.4.2
tzdata                                  2026.1
urllib3                                 2.6.3
uuid_utils                              0.14.1
webencodings                            0.5.1
wrapt                                   2.1.2
xlrd                                    2.0.2
xxhash                                  3.6.0
yarl                                    1.23.0
zope.interface                          8.2
zstandard                               0.25.0